+++ lms/src/main/java/com/scit/lms/service/QuestionService.java
@@ -24,7 +24,7 @@ public interface QuestionService {
     ArrayList<Option> selectOptions(long qid);
 
     // 문제 하나 조회
-    Question selectOneQuestion(int qid);
+    Question selectOneQuestion(long qid);
 
     int deleteQuestion(int testid);
